Add files via upload
Welcome to my repository dedicated to outbreak analytics and applied modelling in R! This collection of files showcases various methodologies and analysis used to understand and visualize infectious disease outbreaks.

In this repository, you'll find a range of scripts and visualizations that explore the dynamics of disease spread, the impact of vaccination strategies, and the evaluation of public health interventions. Each file contains valuable insights derived from real-world data, highlighting R's powerful statistical and graphical capabilities.